+++ test/test_llm_output.py
@@ -15,7 +15,7 @@ from pathlib import Path
 # Add parent directory to path to enable imports
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
 
-from event_labelling.CodeStructure_Branching.code_structure_and_branching import assess_branch_meaningfulness
+from event_labelling.CodeStructure_Branching.main import assess_branch_meaningfulness
 
 
 def test_branch_assessment():
